organize components and add styling to all
