+++ package.json
@@ -12,6 +12,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "identity-obj-proxy": "^3.0.0",
     "jest-dom": "^4.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
@@ -29,14 +30,16 @@
     "@types/react-dom": "^18.2.4",
     "babel-jest": "^29.7.0",
     "babel-loader": "^8.3.0",
-    "css-loader": "^6.7.3",
+    "css-loader": "^6.9.0",
     "file-loader": "^6.2.0",
     "html-loader": "^4.2.0",
     "html-webpack-plugin": "^5.5.1",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "react-test-renderer": "^18.2.0",
-    "style-loader": "^3.3.2",
+    "sass": "^1.69.7",
+    "sass-loader": "^14.0.0",
+    "style-loader": "^3.3.4",
     "ts-jest": "^29.1.1",
     "ts-loader": "^9.4.2",
     "typescript": "^4.9.5",
@@ -50,6 +53,11 @@
     "testEnvironment": "jsdom",
     "transform": {
       "^.+\\.(js|jsx|ts|tsx)$": "babel-jest"
+    },
+
+    "moduleNameMapper": {
+      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
+      "\\.(css|less|scss|sass)$": "identity-obj-proxy"
     }
   }
 }
