+++ public/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
     <head>
-        <title>Template Site</title>
+        <title>Stopwatch - Jack Stasiulis</title>
     </head>
     <body>
         <section id='root'></section>
