+++ src/App.css
@@ -1,0 +1,20 @@
+* {
+    box-sizing: border-box;
+    margin: 0;
+    padding: 0;
+    font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
+    user-select: none;
+}
+
+#root{
+    height: 100vh;
+    background-color: #e6e2db;
+    display: flex;    
+    justify-content: center;
+}
+
+.app {
+    display: flex;
+    align-items: center;
+}
+
