+++ src/Components/StopWatch/StopWatch.css
@@ -1,0 +1,45 @@
+.stopwatch{
+    display: flex;
+}
+
+.stopwatch__title{
+    font-size: 3.5rem;
+}
+
+.stopwatch__main-time{
+    display: flex;
+    justify-content: center;
+    font-size: 5rem;
+    font-family: monospace; 
+}
+
+.stopwatch__buttons{
+    display: flex;
+    justify-content: center;
+}
+
+.stopwatch__lap-list{
+    list-style-type: none;
+    display: flex;
+    flex-direction: column;
+    justify-content: left;
+
+    /* add scrolling to lap list */
+    max-height: 12rem;
+    overflow-y: auto;
+}
+
+.stopwatch__lap-list--container{
+    display: flex;
+    flex-direction: column;
+    margin: 0 0 0 3rem;
+}
+
+.stopwatch__lap-list--title{
+    font-size: 2rem;
+}
+
+.stopwatch__lap-list--item{
+    width: 4rem;
+    padding: 0.2rem 0;
+}
