+++ src/Components/StopWatchButton/StopWatchButton.css
@@ -1,0 +1,19 @@
+.stopwatch__button{
+    width: 8rem;
+    height: 3rem;
+    border-radius: 0;
+    user-select: none;
+    border: 1px solid black;
+    background-color: #e6e2db;
+    &:hover{
+        background-color: #d8d4cd;
+    }
+    &:active{
+        background-color: #cac6be;
+    }
+}
+
+.stopwatch__button--container:nth-child(2) .stopwatch__button {
+    border-left: none;
+    border-right: none;
+  }
